Move s[n]printf to lib for release
