+++ Makefile
@@ -139,6 +139,7 @@ C_SRCS += lib/strncpy.c
 C_SRCS += lib/crypto/tinycrypt/hmac.c
 C_SRCS += lib/crypto/tinycrypt/sha256.c
 C_SRCS += lib/crypto/hkdf.c
+C_SRCS += lib/sprintf.c
 C_SRCS += common/hv_main.c
 C_SRCS += common/hypercall.c
 C_SRCS += common/schedule.c
