+++ include/hv_lib.h
@@ -42,5 +42,6 @@
 #include <list.h>
 #include <bits.h>
 #include <strtol.h>
+#include <sprintf.h>
 
 #endif /* HV_LIB_H */
